Domain Event Logging

+++ src/service/ToolService.kt
@@ -1,56 +1,75 @@
 package cloud.fabx.service
 
+import cloud.fabx.application.AuthorizationException
+import cloud.fabx.application.XPrincipal
 import cloud.fabx.db.DbHandler.dbQuery
+import cloud.fabx.domainEvent
 import cloud.fabx.dto.EditToolDto
 import cloud.fabx.dto.NewToolDto
 import cloud.fabx.dto.ToolDto
+import cloud.fabx.logger
 import cloud.fabx.model.Device
+import cloud.fabx.model.Mapper
 import cloud.fabx.model.Qualification
 import cloud.fabx.model.Tool
-import cloud.fabx.qualificationService
+import net.logstash.logback.argument.StructuredArguments
 import org.jetbrains.exposed.sql.SizedCollection
 
-class ToolService {
+class ToolService(private val mapper: Mapper) {
 
-    suspend fun getAllTools(): List<ToolDto> = dbQuery {
-        Tool.all().map{ toToolDto(it) }.toCollection(ArrayList())
+    private val log = logger()
+
+    suspend fun getAllTools(principal: XPrincipal): List<ToolDto> = dbQuery {
+        principal.requirePermission("get all tools", XPrincipal::allowedToGetAllTools)
+        Tool.all().map { mapper.toToolDto(it) }.toCollection(ArrayList())
     }
 
-    suspend fun getToolById(id: Int): ToolDto? = dbQuery {
-        Tool.findById(id)?.let { toToolDto(it) }
+    suspend fun getToolById(id: Int, principal: XPrincipal): ToolDto? = dbQuery {
+        principal.requirePermission("get tool by id", XPrincipal::allowedToGetTool)
+        Tool.findById(id)?.let { mapper.toToolDto(it) }
     }
 
-    suspend fun createNewTool(tool: NewToolDto): ToolDto = dbQuery {
-        val deviceInDb = Device.findById(tool.deviceId) ?: throw IllegalArgumentException("Could not find deviceId.")
+    suspend fun createNewTool(tool: NewToolDto, principal: XPrincipal): ToolDto = dbQuery {
+        principal.requirePermission("create new tool", XPrincipal::allowedToCreateTool)
+
+        val deviceInDb = Device.findById(tool.deviceId)
+        requireNotNull(deviceInDb) { "Device with id ${tool.deviceId} does not exist" }
 
         val qualificationsInDb: List<Qualification> = tool.qualifications.map {
             val qualification = Qualification.findById(it)
-            requireNotNull(qualification) { "Could not find qualification with id $it" }
+            requireNotNull(qualification) { "Qualification with id $it does not exist" }
 
             qualification
         }
 
-        deviceInDb.let { deviceIt ->
-            val newTool = Tool.new {
-                device = deviceIt
-                name = tool.name
-                pin = tool.pin
-                toolType = tool.toolType
-                toolState = tool.toolState
-                wikiLink = tool.wikiLink
-            }
-
-            newTool.qualifications = SizedCollection(qualificationsInDb)
-
-            toToolDto(newTool)
+        val newTool = Tool.new {
+            device = deviceInDb
+            name = tool.name
+            pin = tool.pin
+            toolType = tool.toolType
+            toolState = tool.toolState
+            wikiLink = tool.wikiLink
         }
+
+        newTool.qualifications = SizedCollection(qualificationsInDb)
+
+        val toolDto = mapper.toToolDto(newTool)
+        log.domainEvent(
+            "new tool: {} by {}",
+            StructuredArguments.keyValue("toolDto", toolDto),
+            StructuredArguments.keyValue("principal", principal)
+        )
+        toolDto
     }
 
-    suspend fun editTool(id: Int, editTool: EditToolDto) = dbQuery {
-        val tool = Tool.findById(id) ?: throw IllegalArgumentException("Tool with id $id does not exist")
+    suspend fun editTool(id: Int, editTool: EditToolDto, principal: XPrincipal) = dbQuery {
+        principal.requirePermission("edit tool", XPrincipal::allowedToEditTool)
 
-        editTool.deviceId?.let {deviceId ->
-            Device.findById(deviceId)?.let {device ->
+        val tool = Tool.findById(id)
+        requireNotNull(tool) { "Tool with id $id does not exist" }
+
+        editTool.deviceId?.let { deviceId ->
+            Device.findById(deviceId)?.let { device ->
                 tool.device = device
             }
         }
@@ -68,23 +87,31 @@ class ToolService {
             }
             tool.qualifications = SizedCollection(qualificationsInDb)
         }
+        log.domainEvent(
+            "edit tool: {} by {}",
+            StructuredArguments.keyValue("toolDto", mapper.toToolDto(tool)),
+            StructuredArguments.keyValue("principal", principal)
+        )
     }
 
-    suspend fun deleteTool(id: Int) = dbQuery {
-        val tool = Tool.findById(id) ?: throw java.lang.IllegalArgumentException("Tool with id $id does not exist")
+    suspend fun deleteTool(id: Int, principal: XPrincipal) = dbQuery {
+        principal.requirePermission("delete tool", XPrincipal::allowedToDeleteTool)
+
+        val tool = Tool.findById(id)
+        requireNotNull(tool) { "Tool with id $id does not exist" }
+
+        log.domainEvent(
+            "delete tool: {} by {}",
+            StructuredArguments.keyValue("toolDto", mapper.toToolDto(tool)),
+            StructuredArguments.keyValue("principal", principal)
+        )
         tool.delete()
     }
 
-    fun toToolDto(tool: Tool): ToolDto {
-        return ToolDto(
-            tool.id.value,
-            tool.device.id.value,
-            tool.name,
-            tool.pin,
-            tool.toolType,
-            tool.toolState,
-            tool.wikiLink,
-            tool.qualifications.map { qualificationService.toQualificationDto(it) }
-        )
+    private fun XPrincipal.requirePermission(description: String, permission: XPrincipal.() -> Boolean) {
+        if (!this.permission()) {
+            log.info("$name tried to $description")
+            throw AuthorizationException("$name not allowed to $description.")
+        }
     }
 }
